backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
openai==1.51.0
httpx==0.27.2
anyio==3.7.1
beautifulsoup4==4.12.3
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.12
lxml==5.3.0
playwright==1.48.0
sentence-transformers==3.2.0
torch>=2.0.0
